Conflicts:
	.gitignore
	MNE/disp3D/inverseview.cpp
	MNE/inverse/rapMusic/pwlrapmusic.cpp
	MNE/inverse/rapMusic/rapmusic.cpp
	MNE/mne/mne_forwardsolution.cpp
	MNE/utils/filterdata.h
	applications/mne_rt_server/connectors/FiffSimulator/fiffsimulator.cpp
	applications/mne_rt_server/mne_rt_server/commandserver.cpp
	applications/mne_x/mne_x/mainwindow.cpp
	applications/mne_x/mne_x/mainwindow.h
	applications/mne_x_libs/mne_x/Interfaces/IPlugin.h
	applications/mne_x_libs/xDisp/newrealtimemultisamplearraywidget.cpp
	applications/mne_x_libs/xDisp/xDisp.pro
	examples/computeInversePwlRapMusic/main.cpp
	examples/computeInverseRapMusic/main.cpp
	examples/examples.pro
	examples/rawClusteredInversePwlRapMusic/main.cpp
	examples/rawClusteredInverseRapMusic/main.cpp
	examples/roiClusteredInversePwlRapMusic/main.cpp
	examples/stClusteredInversePwlRapMusic/main.cpp

+++ applications/mne_rt_server/connectors/Neuromag/shmemsocket.cpp
@@ -139,7 +139,7 @@ int ShmemSocket::receive_tag (FiffTag::SPtr& p_pTag)
 
     /* A sanity check to survive at least some crazy messages */
 
-    if (mess.kind > 20000 || mess.size > (size_t) 100000000)
+    if (mess.kind > 20000 || (unsigned long) mess.size > (size_t) 100000000)
     {
         printf("ALERT: Unreasonable data received, skipping! (size=%d)(kind=%d)", mess.kind, mess.size);
         //dacq_log("ALERT: Unreasonable data received, skipping! (size=%d)(kind=%d)", mess.kind, mess.size);
@@ -151,14 +151,14 @@ int ShmemSocket::receive_tag (FiffTag::SPtr& p_pTag)
     p_pTag->type = mess.type;
     p_pTag->next = 0;
 
-    if (mess.size > (size_t) 0)
+    if ((unsigned long) mess.size > (size_t) 0)
     {
         p_pTag->resize(mess.size);
     }
 
 //    qDebug() << mess.loc << " " << mess.size << " " << mess.shmem_buf << " " << mess.shmem_loc;
 
-    if (mess.loc < 0 && mess.size > (size_t) 0 && mess.shmem_buf < 0 && mess.shmem_loc < 0)
+    if (mess.loc < 0 && (unsigned long) mess.size > (size_t) 0 && mess.shmem_buf < 0 && mess.shmem_loc < 0)
     {
         fromlen = sizeof(from);
         rlen = recvfrom(m_iShmemSock, (void *)p_pTag->data(), mess.size, 0, (sockaddr *)(&from), &fromlen);
@@ -172,7 +172,7 @@ int ShmemSocket::receive_tag (FiffTag::SPtr& p_pTag)
 //        if (mess.type == FIFFT_STRING)
 //            data[mess.size] = '\0';
     }
-    else if (mess.size > (size_t) 0) {
+    else if ((unsigned long) mess.size > (size_t) 0) {
         /*
          * Copy data from shared memory
          */
